STUDY-BUDDY/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
python-dotenv==1.0.1
groq
PyPDF2
pinecone
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.14.1
langchain
langgraph
requests
pydantic
